requirements.txt
functions-framework==3.*
pandas==2.0.*
catboost==1.2
gcsfs==2023.6.0
fastparquet==2023.7.0
pyarrow==12.0.1
gcloud==0.18.3
google-cloud-storage==2.10.0